HAAR-1399: ⬆️ update to spring boot 3

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/externalusersapi/repository/UserRoleRepository.kt
@@ -8,7 +8,7 @@ import java.util.UUID
 @Repository
 class UserRoleRepository(private val databaseClient: DatabaseClient) {
 
-  suspend fun deleteUserRole(userId: UUID, roleId: UUID): Int {
+  suspend fun deleteUserRole(userId: UUID, roleId: UUID): Long {
     val sql = "delete from user_role " +
       "where " +
       "user_id = :userId " +
@@ -22,7 +22,7 @@ class UserRoleRepository(private val databaseClient: DatabaseClient) {
       .awaitFirst()
   }
 
-  suspend fun insertUserRole(userId: UUID, roleId: UUID): Int =
+  suspend fun insertUserRole(userId: UUID, roleId: UUID): Long =
     databaseClient
       .sql("INSERT INTO user_role VALUES (:roleId, :userId)")
       .bind("userId", userId)
